requirements.txt

# Minimal requirements for Auto_Roster (pruned to packages used by the project)
Flask==3.1.1
requests==2.32.4
python-dotenv==1.2.1
line-bot-sdk==3.19.1
google-cloud-vision==3.10.2
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
opencv-python==4.12.0.88
numpy==2.2.6

